Use atom to fetch gas price strategy
